# backend/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "One analysis: which destinations, over which window, ranked how.",
        "examples": [
          {
            "destination_type": "peak",
            "end_datetime": "2026-08-03T02:00:00Z",
            "limit": 5,
            "polygon": {
              "coordinates": [
                [
                  [
                    -122.03,
                    47.44
                  ],
                  [
                    -121.91,
                    47.44
                  ],
                  [
                    -121.91,
                    47.53
                  ],
                  [
                    -122.03,
                    47.53
                  ],
                  [
                    -122.03,
                    47.44
                  ]
                ]
              ],
              "type": "Polygon"
            },
            "sort_by": "precip_total_in",
            "sort_desc": false,
            "start_datetime": "2026-08-01T14:00:00Z"
          }
        ],
        "properties": {
          "custom_destinations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CustomDestination"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Your own destinations, merged into whatever the polygon discovers. A custom row matching a discovered one by name or by coordinates to five decimals replaces it.",
            "title": "Custom Destinations"
          },
          "destination_type": {
            "$ref": "#/components/schemas/DestinationType",
            "description": "What to discover inside the polygon. `GET /api/capabilities` lists the types this deployment actually supports, which is narrower than this enum."
          },
          "end_datetime": {
            "description": "End of the forecast window, ISO 8601, inclusive of the hour it lands in. Must not precede `start_datetime`.",
            "format": "date-time",
            "title": "End Datetime",
            "type": "string"
          },
          "limit": {
            "default": 10,
            "description": "How many ranked rows to return. Discovery is never sampled, so this trims the response, not the work: every candidate is forecast and ranked before the cut.",
            "title": "Limit",
            "type": "integer"
          },
          "max_elevation_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Drop candidates above this elevation.",
            "title": "Max Elevation Ft"
          },
          "min_elevation_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Drop candidates below this elevation. Candidates with an unknown elevation always pass through rather than being silently dropped.",
            "title": "Min Elevation Ft"
          },
          "polygon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPolygon"
              },
              {
                "type": "null"
              }
            ],
            "description": "Search area for destination discovery. Required unless `destination_type` is `custom`, in which case discovery is skipped entirely and only `custom_destinations` are analyzed."
          },
          "sort_by": {
            "$ref": "#/components/schemas/SortBy",
            "default": "precip_total_in",
            "description": "Metric the ranking sorts on."
          },
          "sort_desc": {
            "default": false,
            "description": "Sort direction. False ranks lowest first, which is the useful default: driest, calmest, coldest, cleanest. True flips it to wettest, windiest, warmest, smokiest.",
            "title": "Sort Desc",
            "type": "boolean"
          },
          "start_datetime": {
            "description": "Start of the forecast window, ISO 8601. A naive timestamp is read as UTC. Setting this equal to `end_datetime` means current conditions: the single hour containing that moment.",
            "format": "date-time",
            "title": "Start Datetime",
            "type": "string"
          }
        },
        "required": [
          "destination_type",
          "start_datetime",
          "end_datetime"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "DestinationType": {
        "enum": [
          "peak",
          "trailhead",
          "lake",
          "custom"
        ],
        "title": "DestinationType",
        "type": "string"
      },
      "ErrorResponse": {
        "description": "Body of a hand-raised API error.\n\nNote that a 422 differs: request validation is Pydantic's, and its `detail`\nis a list of per-field objects rather than a string. That shape is\ndocumented separately as `HTTPValidationError`.",
        "properties": {
          "detail": {
            "description": "Plain-language explanation of what went wrong, written to be shown to an end user unmodified.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SortBy": {
        "enum": [
          "precip_total_in",
          "precip_max_in_hr",
          "wind_avg_mph",
          "wind_max_mph",
          "temp_min_f",
          "temp_avg_f",
          "temp_max_f",
          "aqi_avg",
          "aqi_max"
        ],
        "title": "SortBy",
        "type": "string"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Bluebird on GitHub",
      "url": "https://github.com/zimmertr/bluebird"
    },
    "description": "\nBluebird ranks outdoor destinations by their forecast weather.\n\nDraw a polygon and every named peak, trailhead, or lake inside it is discovered\nfrom OpenStreetMap, given a real hourly forecast, and ranked. Or skip discovery\nentirely and send your own coordinates.\n\nThere is no API key, no authentication, and no rate limit. Every upstream is\nfree and keyless, so please be considerate: a single analysis can fan out to\nhundreds of forecast requests.\n\nStart with `GET /api/capabilities` to learn the limits this deployment enforces,\nand `GET /api/version` to confirm which build you are talking to.\n",
    "license": {
      "name": "GPL-3.0-only",
      "url": "https://www.gnu.org/licenses/gpl-3.0-standalone.html"
    },
    "summary": "Map-based weather window finder for hikers and mountaineers.",
    "title": "Bluebird",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze": {
      "post": {
        "description": "Discovers every named destination of the requested type inside the polygon, unions in any `custom_destinations`, fetches a real hourly forecast for each, and returns the top `limit` ranked by `sort_by`.\n\nDiscovery is never sampled, so cost scales with how many destinations the polygon contains, not with `limit`. A large polygon over dense terrain can take tens of seconds. For a progress feed instead of a single long wait, use `POST /api/analyze/stream`.",
        "operationId": "analyze_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request parsed but does not describe a runnable analysis: the window is inverted, the destination type is not discoverable, `custom_destinations` is missing for a custom analysis, the elevation band excludes every candidate, or the candidate count exceeds the cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An upstream failed: every Overpass mirror was unreachable, or the weather API did not answer. Transient and worth retrying. Air quality is exempt, since a failure there degrades to null rather than failing the analysis."
          }
        },
        "summary": "Rank destinations by forecast weather",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analyze/stream": {
      "post": {
        "description": "Identical analysis to `POST /api/analyze`, delivered as Server-Sent Events so a caller can show progress instead of waiting on one long request.\n\nThe HTTP status is **always 200**, including for failures: the connection is already open and streaming by the time most problems surface. Check the event stream, not the status code. Four event types arrive as `data:` lines carrying a JSON object with a `type` field:\n\n- `status` \u2014 a human-readable phase message in `message`\n- `progress` \u2014 `processed`, `total`, and `percent` counters\n- `result` \u2014 the terminal success event, carrying a full `AnalyzeResponse` in `data`\n- `error` \u2014 the terminal failure event, with the reason in `message`\n\nExactly one `result` or one `error` ends the stream.",
        "operationId": "analyze_stream_api_analyze_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "example": "data: {\"type\": \"status\", \"message\": \"Searching for Destinations\u2026\"}\n\ndata: {\"type\": \"progress\", \"processed\": 50, \"total\": 120, \"percent\": 41}\n\ndata: {\"type\": \"result\", \"data\": {\"results\": [], \"total_queried\": 0}}\n\n",
                  "type": "string"
                }
              }
            },
            "description": "An SSE stream. Ends with either a `result` or an `error` event."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rank destinations, streaming progress as it goes",
        "tags": [
          "analysis"
        ]
      }
    },
...
  },
...
}